pyproject.toml
[tool.poetry]
name = "hamilton"
version = "0.1.0"
description = ""
authors = ["zilto"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sf-hamilton = "^1.22.4"
networkx = "^3.1"
websockets = "^11.0.2"
ipykernel = "^6.23.0"
pygls = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
